Refs #7: Use new CloudFormation support for Lambda@Edge
Rather than allowing the CloudFront distribution to be deployed by CloudFormation and then
updating the distribution manually ourselves to add the Lambda association (as was done in
v1 of this plugin), we can now use the CloudFormation built-in support for Lambda@Edge.

The plugin is still needed because CloudFront (via CloudFormation) needs a specific
_version_ of a Lambda function. Thus, in our CloudFormation template (a.k.a.
serverless.yml in this context), we would need to add the function associations to our
distribution by adding `{ Ref: 'TheFunctionVersionLogicalID' }`. Unfortunately, though,
the function version is named "randomly" by Serverless (actually, the logical name is
based on the SHA checksum of the code zip), so we can't compute the value ourselves in the
template.

Thus, the plugin reads our event triggers from the serverless.yml function configuration
and then programmatically updates the CloudFormation template to add the Lambda function
associations to the versions of the function that are being deployed.